+++ qiskit/transpiler/exceptions.py
@@ -17,3 +17,29 @@ class TranspilerError(QiskitError):
 
 class TranspilerAccessError(QiskitError):
     """Exception of access error in the transpiler passes."""
+
+
+class CouplingError(QiskitError):
+    """Base class for errors raised by the coupling graph object."""
+
+    def __init__(self, *msg):
+        """Set the error message."""
+        super().__init__(*msg)
+        self.msg = ' '.join(msg)
+
+    def __str__(self):
+        """Return the message."""
+        return repr(self.msg)
+
+
+class LayoutError(QiskitError):
+    """Errors raised by the layout object."""
+
+    def __init__(self, *msg):
+        """Set the error message."""
+        super().__init__(*msg)
+        self.msg = ' '.join(msg)
+
+    def __str__(self):
+        """Return the message."""
+        return repr(self.msg)
