+++ qiskit/transpiler/passes/mapping/cython/stochastic_swap/utils.pyx
@@ -13,7 +13,7 @@ import numpy as np
 from libc.stdlib cimport calloc, free
 from libcpp.vector cimport vector
 
-from qiskit.mapper.layout import Layout
+from qiskit.transpiler.layout import Layout
 
 
 cdef class EdgeCollection:
